Refactor index path selectors
