+++ packages/noya-state/src/selectors/indexPathSelectors.ts
@@ -1,0 +1,57 @@
+import type Sketch from '@sketch-hq/sketch-file-format-ts';
+import { IndexPath } from 'tree-visit';
+import { ApplicationState, Layers } from '../index';
+import { findIndexPath, INCLUDE_AND_SKIP } from '../layers';
+import type { UUID } from '../types';
+import { getCurrentPage, getCurrentPageIndex } from './pageSelectors';
+
+export type LayerIndexPath = { pageIndex: number; indexPath: IndexPath };
+
+export type LayerIndexPaths = { pageIndex: number; indexPaths: IndexPath[] };
+
+export const getLayerIndexPath = (
+  state: ApplicationState,
+  id: UUID,
+): LayerIndexPath | undefined => {
+  const page = getCurrentPage(state);
+  const pageIndex = getCurrentPageIndex(state);
+  const indexPath = findIndexPath(page, (layer) => layer.do_objectID === id);
+
+  return indexPath ? { pageIndex, indexPath } : undefined;
+};
+
+export const findPageLayerIndexPaths = (
+  state: ApplicationState,
+  predicate: (layer: Sketch.AnyLayer) => boolean,
+): LayerIndexPaths[] => {
+  return state.sketch.pages.map((page, pageIndex) => ({
+    pageIndex: pageIndex,
+    indexPaths: Layers.findAllIndexPaths(page, predicate),
+  }));
+};
+
+export const getSelectedLayerIndexPaths = (
+  state: ApplicationState,
+): IndexPath[] => {
+  const page = getCurrentPage(state);
+
+  return Layers.findAllIndexPaths(page, (layer) =>
+    state.selectedObjects.includes(layer.do_objectID),
+  );
+};
+
+export const getSelectedLayerIndexPathsExcludingDescendants = (
+  state: ApplicationState,
+): IndexPath[] => {
+  const page = getCurrentPage(state);
+
+  return Layers.findAllIndexPaths<Sketch.AnyLayer>(page, (layer) => {
+    const included = state.selectedObjects.includes(layer.do_objectID);
+
+    if (included && layer._class === 'artboard') {
+      return INCLUDE_AND_SKIP;
+    }
+
+    return included;
+  });
+};
